+++ composeApp/src/commonMain/kotlin/com/cowork/app_client/util/TimeUtils.kt
@@ -1,3 +1,3 @@
 package com.cowork.app_client.util
 
-internal expect fun nowPlusHoursIso8601(hours: Int): String
+internal expect fun nowPlusHoursIso8601(hours: Double): String
