+++ composeApp/src/jvmMain/kotlin/com/cowork/app_client/util/TimeUtils.kt
@@ -1,4 +1,4 @@
 package com.cowork.app_client.util
 
-internal actual fun nowPlusHoursIso8601(hours: Int): String =
-    java.time.Instant.now().plusSeconds(hours * 3600L).toString()
+internal actual fun nowPlusHoursIso8601(hours: Double): String =
+    java.time.Instant.now().plusSeconds((hours * 3600).toLong()).toString()
